[utilities] Combine two algorithms for better efficiency
The existing algorithm is fast on small trees. The new algorithm is
faster on larger trees.
